SL-3829 Refactor VersionedLinkedInClient (#249)
